+++ Sentral/SentralLibrary/Database/DataClasses/UserDetailedData.cs
@@ -4,10 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace SentralLibrary.DataClasses;
-public class UserSimpleData
+namespace SentralLibrary.Database.DataClasses;
+public class UserDetailedData
 {
-    public string? CardID { get; set; }
+    private readonly Random random = new();
+
     public string? FirstName
     {
         get => firstName;
@@ -30,7 +31,25 @@ public class UserSimpleData
             }
         }
     }
+    public string? Email { get; set; }
+    public string? CardID { get; set; }
+    public string? CardPin { get; set; }
+    public DateTime StartValidityTime { get; set; }
+    public DateTime EndValidityTime { get; set; }
 
     private string? firstName;
     private string? lastName;
+
+    public UserDetailedData()
+    {
+        CardPin = GeneratePin();
+        StartValidityTime = DateTime.MinValue;
+        EndValidityTime = DateTime.MinValue;
+    }
+
+    private string GeneratePin()
+    {
+        int number = random.Next(0, 10000);
+        return number.ToString("D4");
+    }
 }
